sambacc/common: create common.py from common parts of main.py
The samba-container command that is based on main.py shares a bunch of
functions with dcmain.py the file backing the samba-dc-container
command.
Move a bunch of that common functionality to common.py a new file
for these common things.
